Improve atomics, task wake, tick usage & tests
Clarify and fix atomic CAS semantics and implementation: update ARM header docs to describe retry behaviour; change RISC-V CAS asm to retry spurious sc.w failures and set the failure flag only on genuine mismatches. Tidy kernel atomic code comments/formatting. Change task wake logic to stop preserving delay_ticks when unlinking (delay budget is recomputed by caller). Prevent 32-bit overflow in os_cpu_usage_get by using a 64-bit multiply/divide. Remove duplicate includes in rp235x soc_cb.c. Improve test bench robustness by discarding absurdly large cycle samples (add TEST_BENCH_MAX_SAMPLE) to avoid poisoning worst-case measurements.

+++ kernel/os_tick.c
@@ -255,7 +255,11 @@ uint32_t os_cpu_usage_get(void)
             idle_ticks = total_ticks;
         }
 
-        usage_percent = ((total_ticks - idle_ticks) * 100U) / total_ticks;
+        /* The busy-tick product runs in 64 bits: a window of more than ~42 million
+         * ticks (about 12 hours at 1 kHz) would otherwise overflow the 32-bit
+         * multiplication and report a wrong percentage. */
+        usage_percent = (uint32_t)((((uint64_t)total_ticks - (uint64_t)idle_ticks) * 100ULL) /
+                                   (uint64_t)total_ticks);
     }
 
     return usage_percent;
